Auto-trigger upscaling if w * h > 1024 * 1024
This should allow higher resolution images to be generated, without
having as high of a risk of running out of VRAM. Navigator will tell the
backend to generate the requested image at the resolution / 2, which the
backend will then upscale back to the original requested resolution. In
theory.

Backend doesn't seem to support width OR height going over 2048, so we
reject image generation requests if they specify a width/height that is
higher than 2048.

At least, it does on the frontend - I haven't checked via the API,
however we're not likely going to be able to generate images past that
resolution even with HR Fix being enabled, but I'll re-evaluate this
later.