Raise the maximum width/height of generated images
We also now are a bit more flexible with the limit, the new limit works
the same way that we determine whether HR-Fix should be forced or not.
That is, we take the width and multiply it by the height, and ensure
that value isn't higher than a 2K image @ a 16:9 Aspect Ratio
(2560x1440).

This is because of the fact that we place limits on the max resolution
due to avoiding running out of VRAM on the backend - the VRAM usage is
going to be affected by the total pixels to generate and work with, thus
its not really a case of "max width" OR "max height", but rather a "max
total pixels" instead makes more sense.