API: Introduce task interrupt endpoint
Requires authentication and specifying job ID - if the current running job is not the requested ID (or the user doesn't own it), then reject. If passed, emits a task-interrupted event, however due to the way things currently work a task-finished event will fire when the backend cancels the job. The upside is that the user still gets to have whatever was generated.