Txt2img: Accept subseed and subseed_strength parameters
These, if present, get passed to the backend - these will allow you to effectively regenerate a similar image with variations to it. Subseed strength is a value between 0-1, 0 being no variation, 1 being extreme variation. The subseed itself is effectively the seed of the variation, generally you'll probably want to pass -1 to have a random one generated by the backend.