Queue: Do not attempt to check progress for a Task if it is not active on backend.
Now we inform the backend of our own Job ID, and we can use a separate endpoint to basically "peek" at the task on the backend. If the Task ID doesn't match what we set, then we do not return any progress updates until the Task ID matches.

Basically, this fixes the following scenario: An admin is running a task on the backend directly, and a end-user submits a request via Navigator. Navigator's queue will not know that a task is already running on the backend, so it will attempt to send off the task to the backend. This is fine, but Navigator would previously return the progress results of the Admin's direct-task which is not desired.

In the above example, with these changes, the "peek" check will fail, and thus we will skip that progress check until it finally passes.