Txt2Img: Implement dedicated upscale-hrf endpoint (Closes #1)
This endpoint allows you to pass in an existing saved job ID, which will then automatically parse the previous generation's parameters and then will perform an upscale with High-Resolution Fix enabled to the backend. It will also pass the previous version of the image to the backend, so that the backend doesn't need to regenerate this before performing the HRF.

The backend software didn't actually properly handle this, despite having the parameter exposed in the API. I patched the software in order to allow passing a base64 string, which it will decode for the generation.

Additionally, some excess debug messages were removed.