API+User: Add image (internal) metadata endpoint
This is separate from the /info endpoint that already is public and doesn't require authentication. This endpoint will allow retrieving the metadata that Navigator internally keeps about images, such as their category ID and the time they were created. This endpoint requires authentication, and will return a 403 if you request metadata for an image you did not create.