API: Paginate the /api/user/jobs endpoint
Works the way most paginated REST APIs do in that calling the route will return a count of all items, the amount of pages, the current page (starts at 1 if page query param isn't passed in), previous/next page, and of course the actual array of images/jobs on that page.

Supports allowing the client to specify the amount of items per page if desired via the limit query param (by default is 10) as well.

As an added bonus, we allow passing in the category_id field as a way to filter the returned jobs server-side instead of the client needing to filter a massive list. Passing in the category_id will dynamically adjust page counts as expected.