API: Refactor HR Fix logic and adjust default denoising behavior
Ensure that the new `hrf_steps` and `denoising_strength` are actually respected when calling the /upscale-hrf/ endpoint.

Updated HR Fix logic to improve clarity and reliability, replacing `force_upscale` with `force_hr_fix` for consistency. Adjusted default `denoising_strength` to 0.35 when HR Fix is enabled and added support for customizable `hrf_steps`. Corrected width/height handling to prevent invalid values when implicit HR Fix is required.

Changed txt2img endpoint to no longer explicit trigger HR Fix when passing in a denoising strength - it will only be used if the `force_hr_fix` parameter is passed, or if HR Fix is implicitly required due to the w/h exceeding 1024*1024 pixels.

Finally, instruct the backend to keep track of the last used model so that when Navigator restarts, it is able to pick up the previous state properly.