API: Fix handling of HR Fix parameters
This commit fixes an issue that caused upscale images to come out blurrier than the base image. The cause of this was that when not passing an "upscaler_name", before sending off the request to the SD API we did not attach a default - which meant no upscaler was used, resulting in the blurry behavior.

Additionally, this commit fixes the handling of clamping `hrf_steps` - the previous method of attaching `.clamp` to the Number prototype didn't actually seem to work in all cases. Moved it to its own standard function and removed the prototype extension.