Refactor queue/socket management and worker logic
Moved queue management and processing logic from `api.js` to a new `queueWorker.js` module for better separation of concerns and maintainability. Refactored related functions and replaced direct access to the queue with new helper methods. Socket management was moved to `socketManager.js`.

Updated `api.js` to use refined `socketManager` and queue utilities.

A sneaky feature added in this refactor was allowing the job progress update interval to be defined by the user via the `JOB_PROGRESS_CHECK_INTERVAL` environmental variable. If not set, it will default to 2500. The `.env.example` file was updated to include this, along with a hefty warning to not set it too low.