API+Interrupt: Enhance task interruption endpoint
Update API to allow interrupt to also act as a dequeue endpoint.

Previously, this endpoint would fail if the provided job ID was not actively running. Now it will remove the Job from the queue if it's not already running and delete the Job database entry.

It will of course still return an error if the user doesn't own the specified Job ID or the Job ID doesn't exist at all (either as the active job or in the queue).