Security: Add roles; Require role to generate Access Tokens.
Introduce role-based access control with new permission roles, user authorization, and disabling functionality. Added admin routes for managing user roles, database migrations for user role storage, and integration with Discord API for validation. Extended middleware for authentication and role enforcement.

Update `.env.example` to specify a `SUPERUSER_ADMIN_TOKEN` - this is a static access token that grants callers to the API the highest privileges. This token gets passed in as an Authorization header, though without `Bearer` or any other prefix.

Additionally, this changeset will enforce that when attempting to log in to the API that the user either has implicit or explicit authorization to do so.

Users will have *implicit* authorization if they already have pre-existing images generated (this can be either via the "Bot"/Non-user facing API or just by having generated images before this changeset). In this case, they'll be added to the users table with the standard role ("Apprentice" - ID of `1`).

Users can also be given explicit authorization via the new `/api/admin/authorize-user/:discordId` endpoint. This endpoint can be used by users who have the administrator role ("ARCHITECT" - ID of `3`) or via the `SUPERUSER_ADMIN_TOKEN`.

In the case of a newly spun up instance of Navigator, the operator should ensure a `SUPERUSER_ADMIN_TOKEN` is set and use this token to call `/api/admin/authorize-user/:discordId` against their own Discord ID and any other members they wish to be able to grant access to.

Subsequently, a user can have their access revoked by an administrator via the `/api/admin/disable-user/:discordId` endpoint which sets their role to 0. Attempting to call an authenticated endpoint while your role is set to 0 will return a 401.

Codebase wise, introduce a AuthenticatedRouter and AdministratorRouter so that future route handlers do not constantly need to do a `if(req.user)` check.