3P: Add 3P ID to the metadata cache table; Add getting models by ID
CivitAI models can now be fetched by the version ID rather than just their hash. Due to this, when grabbing the model by the hash, save a copy of the 3P (CivitAI in this case) ID so that it can easily be referenced in the cache later on by getModelVersionById.

As part of the database migrations for this, reset all cache times back to 1/1/2000 to invalidate their entries. This will allow model metadata to be rediscovered with their (version) IDs added to the database.