LoRA+API: Fall back to SHA256 if the Forge API doesn't return a hash
This can only be done if the operator specifies where Forge stores the model files. If this is not specified, fall back to the previous behavior (ignore the LoRA model).

Since calculating hashes can be an expensive operation (especially if being done remotely, such as over SSHFS...) we store the calculated hashes in the database.

Then, if we do have a SHA256 hash, use that when passing to CivitAI's model lookup API which will work in-place of the AutoV3 hash that Forge normally returns. These results still integrate with our caching system for this, so that we're not unnecessarily calling CivitAI's API.