3P: Overhaul CivitAI model download endpoint
Now allows for downloading of Forge-supported model types (or rather what I know to be supported).

This API endpoint now requires the user to provide the model version ID of the model they want to download (instead of just the "base" model ID), so that specific versions can be downloaded.

Added authentication to this route (requires Artificer or higher).

Instead of downloading the first found model file, download the one that is tagged as "primary" by the CivitAI API (most of the time these are both the same, but this isn't always the case).

Before downloading models, check to see if it is already downloaded - this is done by checking if the filename of the requested model/version exists in the directory associated with the type of the model. If it does exist already, return a 400.

Finally, after a model has been downloaded, send a reload signal to Forge so that the model can actually be used.

Note that unfortunately, this API endpoint still doesn't really have a way to return progress. The call will just hang until the download is complete (either with a success code or an error code if it fails). Theoretically, we can probably implement progress via sending it over the WebSocket.