3P: Remove isArtificer guard from downloads_enabled route
This causes issues with Compass since it tries to hit this endpoint to see if it should display the navigation item - if the user doesn't have the role, it returns a 401 which causes Compass to think the user's session has expired.

Should be fine to keep removed since it alone doesn't allow users to trigger downloads.