API+LoRAs: Implement local override support for LoRAs
Support for this comes in two formats:

- You can add a database entry with provider type 0 with metadata (see `types/models.ts#LocalOverrideLora` for structure)
- Alternatively, if you have a copy of the CivitAiModelVersion format (or feel very confident about "spoofing" it - not recommended), you can insert it into the database and set `updates_disabled` for that row to true so that it will not get overwritten (and will not try to ask CivitAI for a new copy once the cache is older than 7 days).

The use case for an override is if you got the LoRA (or other model type, later on) from somewhere outside CivitAI or the model is no longer hosted there. This means if you have your own trained models, or they're downloaded from HuggingFace, TensorArt, etc. they can still have decent metadata returned by the API for downstream clients.

If a SHA256 sum for a model cannot be calculated (and a hash identifier isn't sent by Forge), it will fall back to using the alias from Forge as a key. This is useful in cases where Navigator and Forge do not have shared filesystem access to the model files, as you can then add an override for it.

The `/api/models/loras` endpoint will now indicate the provider type (0 for a local override, 1 for CivitAI) which clients can use to determine what data to render.

The `/api/models/loras/override/:hash` endpoint can be used to add local data into the database. This endpoint requires Administrator access or higher.

When the cache entry for a LoRA's entry expires, upon a failure to retrieve the metadata from CivitAI, return the cached data instead of nothing - this could happen in the instance of a model being taken off the API.

(A warning is printed in the console about this, it is advised to look into this - if the author took down their model, for example, please be respectful and do not redistribute images created with it or redistribute the model/metadata)

Note that due to the relocation of how caching is handled (a more abstract location rather than being directly associated with the CivitAI class), models downloaded from CivitAI through the Navigator API will no longer pre-cache the metadata. Instead, it will be cached the next time it is retrieved from the LoRAs endpoint - this should be updated later on.