Security: Explicit grant APEX role for the first (active) user in the database
Added logic to automatically assign the APEX (Superuser) role to the first user added to the users table. Introduced a helper function to check if the table is empty before setting roles.

Note that since this endpoint still requires Administrator ("ARCHITECT") authorization, to call the endpoint without having any active Administrator accounts you'll need to pass the `SUPERUSER_ADMIN_TOKEN` in the `Authorization` header for the request body (do not pass the `Bearer` prefix).