Auth: Add support for enabling a demo user account
Introduce a configurable `ENABLE_DEMO_USER` environment variable to manage access for a demo user account. Disabled by default, even if the variable is undefined.

To authorize as a demo user, use the `DEMO_USER` authorization code against the `/api/auth/login` endpoint. If enabled, this will return a valid JWT token which can be used for most authenticated endpoints, along with a fake Discord user object. If disabled (default), attempting to use this authorization code results in a 403.

The demo account is intended more for showcasing purposes of Navigator & downstream clients - but is included in case other operators want to use it as well.

Demo users receive the "Artificer" permission role which allows them to use the model download API endpoint and temporary JWT tokens. These tokens always have an expiration of 3 hours (even if the operator has adjusted token expiration times).

If the demo account is disabled, demo-related authentication requests are rejected, even if the token is still valid.

Note that you need to restart Navigator if toggling the variable (as with all other variables). Additionally, images created by the demo account are persisted even if the demo account is disabled - they need to be deleted either via the UI or dropped from the database manually for now.