Auth+Security: Update auth payload; Update auth endpoint guard
Upon a successful authentication via the `/api/auth/login` endpoint, the response payload now explicitly contains the user's `role` rather than needing to call `/api/user/me` to retrieve it.

Update the `isAuthenticated` route guard so that if a user with a disabled account hits an endpoint while having a valid JWT they will receive a 403 response (instead of 401). This was done to replicate the same behavior as when said users hit the `/api/auth/login` endpoint so that clients have consistent responses.