Meta: Setup Prettier and format codebase
Navigator has had some consistency issues regarding the codebase,
especially regarding whether semicolons are (or are not) used at the end
of statements. Setup Prettier to enforce a consistent style.

While I'm not the biggest fan of some of Prettier's opinionated choices
of style, it is at least *consistent*.

This also messes with the repo's `blame`-ability but oh well. I would rather
it happens now instead of later on down the line when Navigator is even
larger.

Update docs to include instructions on setting up a pre-commit hook for
Prettier so that even when using an editor that doesn't support
Prettier, the style choices will still be enforced.