API: Update scheduler name validation logic
Slight tweak, whenever the scheduler does not validate, try it again
with the scheduler in all lowercase. This is because Forge's info API
returns the scheduler's "label" whereas Forge requires the "name" to be
sent in the txt2img API.

This resulted in upscaling using the wrong scheduler because when we
retrieve the metadata of the image, we get back the "label" for the
scheduler thus not passing validation and getting set back to
"automatic". Generally this doesn't cause *too* many issues, but when
using Flux it completely breaks upscaling (Flux deosn't support the
"automatic" scheduler).